# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ahmed Riaz Textile Factory website and content management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SiteSettings {
  brandName: string;
  location: string;
  phone: string;
  whatsapp: string;
  email: string;
  officeAddress: string;
  factoryAddress: string;
  businessHours: string;
  whatsappMessage: string;
  heroImage?: string;
  /** @nullable */
  heroVideo?: string | null;
}

export interface SiteSettingsUpdate {
  brandName?: string;
  location?: string;
  phone?: string;
  whatsapp?: string;
  email?: string;
  officeAddress?: string;
  factoryAddress?: string;
  businessHours?: string;
  whatsappMessage?: string;
  heroImage?: string;
  /** @nullable */
  heroVideo?: string | null;
}

export interface FactoryStat {
  value: string;
  label: string;
  displayOrder: number;
}

export type ContentCollection = typeof ContentCollection[keyof typeof ContentCollection];


export const ContentCollection = {
  designs: 'designs',
  products: 'products',
  machinery: 'machinery',
  services: 'services',
  blogs: 'blogs',
  faqs: 'faqs',
  timeline: 'timeline',
  categories: 'categories',
} as const;

export type ContentRecordMeta = { [key: string]: unknown };

export interface ContentRecord {
  id: number;
  collection: ContentCollection;
  slug: string;
  title: string;
  shortDescription: string;
  description: string;
  category: string;
  image: string;
  images: string[];
  /** @nullable */
  video: string | null;
  published: boolean;
  featured: boolean;
  displayOrder: number;
  meta: ContentRecordMeta;
}

export type ContentInputMeta = { [key: string]: unknown };

export interface ContentInput {
  collection: ContentCollection;
  slug?: string;
  title: string;
  shortDescription: string;
  description: string;
  category?: string;
  image?: string;
  images?: string[];
  /** @nullable */
  video?: string | null;
  published?: boolean;
  featured?: boolean;
  displayOrder?: number;
  meta?: ContentInputMeta;
}

export type ContentUpdateMeta = { [key: string]: unknown };

export interface ContentUpdate {
  slug?: string;
  title?: string;
  shortDescription?: string;
  description?: string;
  category?: string;
  image?: string;
  images?: string[];
  /** @nullable */
  video?: string | null;
  published?: boolean;
  featured?: boolean;
  displayOrder?: number;
  meta?: ContentUpdateMeta;
}

export interface SiteContent {
  settings: SiteSettings;
  stats: FactoryStat[];
  designs: ContentRecord[];
  products: ContentRecord[];
  machinery: ContentRecord[];
  services: ContentRecord[];
  blogs: ContentRecord[];
  faqs: ContentRecord[];
  timeline: ContentRecord[];
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthSession {
  authenticated: boolean;
  /** @nullable */
  username?: string | null;
}

export type DashboardSummaryCounts = {[key: string]: number};

export interface DashboardSummary {
  counts: DashboardSummaryCounts;
  recent: ContentRecord[];
}

export type ListAdminContentParams = {
collection: ContentCollection;
};