# lib/api-zod/src/generated/types/contentCollection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ahmed Riaz Textile Factory website and content management API
 * OpenAPI spec version: 0.1.0
 */

export type ContentCollection = typeof ContentCollection[keyof typeof ContentCollection];


export const ContentCollection = {
  designs: 'designs',
  products: 'products',
  machinery: 'machinery',
  services: 'services',
  blogs: 'blogs',
  faqs: 'faqs',
  timeline: 'timeline',
  categories: 'categories',
} as const;