# lib/api-zod/src/generated/types/factoryStat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ahmed Riaz Textile Factory website and content management API
 * OpenAPI spec version: 0.1.0
 */

export interface FactoryStat {
  value: string;
  label: string;
  displayOrder: number;
}